Move Empty component to reuse
