+++ src/components/index.js
@@ -10,6 +10,7 @@ export { default as PlanPicker } from './planPicker/PlanPicker';
 export { default as SparkPost } from './sparkPost/SparkPost';
 export { default as UsageReport } from './usageReport/UsageReport';
 export { default as ApiKeySuccessBanner } from './apiKeyBanner/ApiKeyBanner';
+export { default as Empty } from './empty/Empty';
 
 export * from './collection';
 export * from './formatters';
